User Validation

+++ src/com/bitlab/entity/Employe.java
@@ -19,12 +19,12 @@ public class Employe
     private String emp_direccion;
     private byte emp_estado;
     private Departament departament;
-    private byte emp_jef_id;
+    private int emp_jef_id;
 
     public Employe() {
     }
 
-    public Employe(int emp_id, String emp_codigo, String emp_nombres, String emp_apellidos, String emp_dui, String emp_nit, String emp_correo, String emp_telefono, double emp_sueldo, String emp_direccion, byte emp_estado, Departament departament, byte emp_jef_id) {
+    public Employe(int emp_id, String emp_codigo, String emp_nombres, String emp_apellidos, String emp_dui, String emp_nit, String emp_correo, String emp_telefono, double emp_sueldo, String emp_direccion, byte emp_estado, Departament departament, int emp_jef_id) {
         this.emp_id = emp_id;
         this.emp_codigo = emp_codigo;
         this.emp_nombres = emp_nombres;
@@ -136,11 +136,11 @@ public class Employe
         this.departament = departament;
     }
 
-    public byte getEmp_jef_id() {
+    public int getEmp_jef_id() {
         return emp_jef_id;
     }
 
-    public void setEmp_jef_id(byte emp_jef_id) {
+    public void setEmp_jef_id(int emp_jef_id) {
         this.emp_jef_id = emp_jef_id;
     }
